+++ lifecycle/data/data_adapter.py
@@ -145,14 +145,9 @@ def serv_instance_is_agent_in_service_instance(service_instance, agent_url):
     return adapter.serv_instance_is_agent_in_service_instance(service_instance, agent_url)
 
 
-# serv_instance_add_agents_to_empty_service_instance:
-def serv_instance_add_agents_to_empty_service_instance(service, user_id, agreement_id, agents_list):
-    return adapter.serv_instance_add_agents_to_empty_service_instance(service, user_id, agreement_id, agents_list)
-
-
-# serv_instance_add_agents_to_empty_service_instance:
-def serv_instance_new_empty_service_instance(service, user_id, agreement_id):
-    return adapter.serv_instance_new_empty_service_instance(service, user_id, agreement_id)
+# serv_instance_replace_service_instance_agents:
+def serv_instance_replace_service_instance_agents(service_instance, service, user_id, sla_template_id, agents_list):
+    return adapter.serv_instance_replace_service_instance_agents(service_instance, service, user_id, sla_template_id, agents_list)
 
 
 ###############################################################################